// src/insights/components/dashboard-visualization/components/TimeSeries2/types.ts
type LineChartDataSeries = { x: Float64Array; y: Float64Array };

type LineChartDataSetData = {
  index: {
    name: string;
    units: string;
    type: "Double" | "Integer";
    dataBase64: string;
  };
  columns: {
    name: string;
    units: string;
    type: "Double" | "Integer";
    dataBase64: string;
  }[];
};

export type ProcessDataSet = (props: {
  dataResult: LineChartDataSetData;
  columnNames: string[];
  dataSetId: string;
  range: Range | null;
}) => LineChartData[] | null;

export type LineChartData = {
  data: LineChartDataSeries;
  label: string;
  min: number;
  max: number;
  dataSetId: string;
};

export type RequestAndPrepareColumnData = (
  params: CommonParams<"TimeSeries"> & {
    dataRange: DataRange | null;
    pointCount: number;
  }
) => Promise<LineChartData[] | null>;

export type LineChartDataWithColor = LineChartData & { color: string };

export type RefreshColumnData<T extends DataVisTypes> = (
  params: CommonParams<T> & {
    width: number;
    range: Range;
  }
) => Promise<LineChartDataWithColor[] | null>;

export type RefreshColumns = (params: {
  range: Range;
  dataSetId: string;
  datasetLabels: string[];
  dashboardId: string;
  visualizationId: string;
  dimensions: Dimensions;
  properties: FilteredVisualizationProperties<TimeSeriesKey>;
}) => Promise<{ chartColumnsWithUpdatedLabels: LineChartDataWithColor[] } | null>;

export type RefreshColmnsHandler = (params: {
  range: Range;
  dataSetId: string;
  datasetLabels: string[];
}) => Promise<{ chartColumnsWithUpdatedLabels: LineChartDataWithColor[] } | null>;

export type FetchColumnData = (
  params: DataSetId & {
    visibleRange: NumberRange | Range | null;
    dashboardId: string;
    visualizationId: string;
    dimensions: Dimensions;
    properties: FilteredVisualizationProperties<TimeSeriesKey>;
    schema: DataSet<"TimeSeries">;
    name: string;
  }
) => Promise<
  | {
      state: "Loaded";
      data: {
        chartProps: {
          xAxisLabel: string;
          yAxisLabel: string;
          xLimits: Range;
          name: string;
        };
        datasetLabelsData: string[];
        chartColumnsWithUpdatedLabels: LineChartDataWithColor[];
        minimapColumns: LineChartDataWithColor[];
      };
    }
  | { state: "Loading" }
  | null
>;
// ...
export type GetColumnData<T extends DataVisTypes> = (
  params: CommonParams<T> & {
    width: number;
    visibleRange: NumberRange | Range | null;
  }
) => Promise<
  | {
      state: "Loaded";
      data: {
        chartColumns: LineChartDataWithColor[];
        xLimits: Range;
        minimapColumns: LineChartDataWithColor[];
      };
    }
  | { state: "Loading" }
  | null
>;